+++ src/Crowdin.Api/Reports/CostEstimationPostEditingGenerateReportRequest.cs
@@ -57,13 +57,13 @@ namespace Crowdin.Api.Reports
             public string? LanguageId { get; set; }
             
             [JsonProperty("fileIds")]
-            public ICollection<int>? FileIds { get; set; }
+            public ICollection<long>? FileIds { get; set; }
             
             [JsonProperty("directoryIds")]
-            public ICollection<int>? DirectoryIds { get; set; }
+            public ICollection<long>? DirectoryIds { get; set; }
             
             [JsonProperty("branchIds")]
-            public ICollection<int>? BranchIds { get; set; }
+            public ICollection<long>? BranchIds { get; set; }
             
             [JsonProperty("dateFrom")]
             public DateTimeOffset? DateFrom { get; set; }
@@ -72,7 +72,7 @@ namespace Crowdin.Api.Reports
             public DateTimeOffset? DateTo { get; set; }
             
             [JsonProperty("labelIds")]
-            public ICollection<int>? LabelIds { get; set; }
+            public ICollection<long>? LabelIds { get; set; }
             
             [JsonProperty("labelIncludeType")]
             public ReportLabelIncludeType? LabelIncludeType { get; set; }
@@ -102,7 +102,7 @@ namespace Crowdin.Api.Reports
             public ICollection<string> LanguageIds { get; set; }
 
             [JsonProperty("userIds")]
-            public ICollection<int> UserIds { get; set; }
+            public ICollection<long> UserIds { get; set; }
 #pragma warning restore CS8618
             
             [JsonProperty("fullTranslation")]
