+++ src/Crowdin.Api/SourceStrings/UploadStringsRequest.cs
@@ -26,7 +26,7 @@ namespace Crowdin.Api.SourceStrings
         public int? ParserVersion { get; set; }
         
         [JsonProperty("labelIds")]
-        public ICollection<int>? LabelIds { get; set; }
+        public ICollection<long>? LabelIds { get; set; }
         
         [JsonProperty("updateStrings")]
         public bool? UpdateStrings { get; set; }
