+++ src/Crowdin.Api/Tasks/TaskCreateForm.cs
@@ -25,13 +25,13 @@ namespace Crowdin.Api.Tasks
 #pragma warning restore CS8618
         
         [JsonProperty("branchIds")]
-        public ICollection<int>? BranchIds { get; set; }
+        public ICollection<long>? BranchIds { get; set; }
         
         [JsonProperty("fileIds")]
-        public ICollection<int>? FileIds { get; set; }
+        public ICollection<long>? FileIds { get; set; }
         
         [JsonProperty("stringIds")]
-        public ICollection<int>? StringIds { get; set; }
+        public ICollection<long>? StringIds { get; set; }
         
         [JsonProperty("type")]
         public TaskType Type { get; set; }
@@ -57,10 +57,10 @@ namespace Crowdin.Api.Tasks
         public bool? IncludePreTranslatedStringsOnly { get; set; }
         
         [JsonProperty("labelIds")]
-        public ICollection<int>? LabelIds { get; set; }
+        public ICollection<long>? LabelIds { get; set; }
         
         [JsonProperty("excludeLabelIds")]
-        public ICollection<int>? ExcludeLabelIds { get; set; }
+        public ICollection<long>? ExcludeLabelIds { get; set; }
         
         [JsonProperty("assignees")]
         public ICollection<TaskAssigneeForm>? Assignees { get; set; }
